fix: address senior review findings (part 1) + plan for remaining work
Demo/CI breakers: CI installs report extra and runs mypy/migrations/build;
compose waits on api health with restart policies; worker loop is crash-safe
with retry backoff; GET /renders/{id} no longer 500s with 2+ designs;
quantity_unit migration has a server default; make infra/env targets work.

Data loss: re-upload supersedes the old photo instead of cascade-deleting
regions and assignments; re-segment soft-deactivates model regions; status
race on /segment fixed; enqueue idempotency race handled.

Estimation: real edge-ratio foreshortening; balcony gets a running length
(no more zero-cost lines); pillar 3-face formula; tiles billed as whole
boxes; geometric-mean user scale; estimate staleness fingerprint with
report 409 when stale; PATCH /measurements only touches sent fields.

Robustness: prod config guard, upload size cap before buffering, CPU work
off the event loop, per-region diffusion holes, XFF-aware rate limiting,
segmentation upsamples argmax not logits, FK-enforced SQLite tests.

Remaining frontend/test/hygiene work is tracked in docs/plans/review-fixes.md.

+++ apps/api/app/providers/render/local.py
@@ -99,7 +99,7 @@ def render_region(
         ).astype(np.float32)
         if r.category == "texture" and r.color_hex:
             tint = _hex(r.color_hex) / 255.0
-            base = base * tint[None, None, :] * 1.15
+            base = (base * tint[None, None, :] * 1.15).astype(np.float32)
     else:
         base = np.empty_like(orig, np.float32)
         base[:] = _hex(r.color_hex)
